Add files via upload
Yup! This is a Rock - Paper - Scissors game. I think it's one of the games that you could play even only based on text and thought I would give it a try. 
To be honest this is fun to play, simple and it makes me wonder how many more games I could make. looking forward to make me some more programs that would make me more intrigued in programming.